packages/hapax-swarm/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hapax-swarm"
version = "0.1.0"
description = "Filesystem-as-bus multi-session coordination library — relay protocol, claim-before-parallel-work, cc-task SSOT for Claude Code swarms."
readme = "README.md"
requires-python = ">=3.12"
license = "LicenseRef-PolyForm-Strict-1.0.0"
license-files = ["LICENSE.txt"]
authors = [
    { name = "Hapax (Oudepode)", email = "hapax@oudepode.studio" },
    { name = "Claude Code", email = "noreply@anthropic.com" },
]
keywords = [
    "swarm",
    "coordination",
    "claude-code",
    "multi-agent",
    "filesystem-as-bus",
    "relay-protocol",
    "research-software",
    "single-operator",
    "infrastructure-as-argument",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/ryanklee/hapax-council"
Repository = "https://github.com/ryanklee/hapax-council"
"Bug Tracker" = "https://github.com/ryanklee/hapax-council/issues"
"Methodology" = "https://hapax.weblog.lol/velocity-report-2026-04-25"
"Citation" = "https://github.com/ryanklee/hapax-council/blob/main/packages/hapax-swarm/CITATION.cff"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hapax_swarm"]

[tool.hatch.build.targets.sdist]
include = [
    "src/hapax_swarm/**",
    "tests/**",
    "README.md",
    "LICENSE.txt",
    "CITATION.cff",
    "pyproject.toml",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "SIM",
    "TCH",
]

[tool.ruff.lint.isort]
known-first-party = ["hapax_swarm"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
